Projeto Finalizado
Projeto concluido, infelizmente não foi possível realizar mais commit's durante o projeto, mas será feito complementações conforme vou tendo mais experiências.

+++ Projeto_Brillare/app/src/main/java/classes/Produto.java
@@ -7,9 +7,9 @@ public class Produto implements Serializable {
     int image;
     String titulo;
     String descricao;
-    int preco;
+    double preco;
 
-    public Produto(int image, String titulo, String descricao, int preco) {
+    public Produto(int image, String titulo, String descricao, double preco) {
         this.image = image;
         this.titulo = titulo;
         this.descricao = descricao;
@@ -28,7 +28,7 @@ public class Produto implements Serializable {
         return descricao;
     }
 
-    public int getPreco() {
+    public double getPreco() {
         return preco;
     }
 }
